+++ src/ts/WebView.ts
@@ -140,6 +140,11 @@ export class WebView {
             this._configuration = configuration;
             HttpKafka.setConfiguration(this._configuration);
 
+            if (!this._configuration.isEnabled()) {
+                const error = new Error('Game with ID ' + this._clientInfo.getGameId() +  ' is not enabled');
+                Diagnostics.trigger('disabled_game', error);
+            }
+
             if(this._configuration.isAnalyticsEnabled()) {
                 if(this._nativeBridge.getPlatform() === Platform.ANDROID) {
                     this._wakeUpManager.setListenAndroidLifecycle(true);
@@ -292,11 +297,17 @@ export class WebView {
 
     private onAdUnitStartProcessed(): void {
         if(this._currentAdUnit) {
+            // A/B test for 2 second refresh after start
+            let magicConstant: number = this._startRefreshMagicConstant;
+            if(this._currentAdUnit.getCampaign().getAbGroup() === 6) {
+                magicConstant = 2000;
+            }
+
             setTimeout(() => {
                 if(!this._mustReinitialize && this._currentAdUnit && this._currentAdUnit.isCached()) {
                     this._campaignRefreshManager.refresh();
                 }
-            }, this._startRefreshMagicConstant);
+            }, magicConstant);
         }
     }